test(updater): extract classifyUpdateError to a pure utility so the test loads
CI failed Unit Tests on the prior push because the test file imported
classifyUpdateError from autoUpdater.ts, which transitively pulls in
app.getPath() and safeHandle IPC registrations that crash at module-
load time in a vitest environment without a real Electron app global.
229 other tests passed; this was the only file that failed to load.

Extracted classifyUpdateError, categorizeDownloadDuration, and
isWindowsRenameLockError to packages/electron/src/main/services/autoUpdaterUtils.ts.
They are pure functions with no Electron, fs, or node-only dependencies.
autoUpdater.ts re-imports and re-exports them for backwards compatibility.
The test now imports from autoUpdaterUtils directly and runs cleanly.

All 8 cases pass locally.
